refactor: rename database query to sql query builder.
